// src/services/types/data.ts
export interface IIngredient {
    _id: string;
    name: string;
    type: string;
    proteins: number;
    fat: number;
    carbohydrates: number;
    calories: number;
    price: number;
    image: string;
    image_mobile: string;
    image_large: string;
    __v: number;
    constructorItemId?: string;
  }

  export interface IIngredientsInitialState {
    burgerIngredientsList: Array<IIngredient>;
    burgerIngredientsListRequest: boolean;
    burgerIngredientsListFailed: boolean;
    burgerIngredientsListFailedText: string | undefined;
  }
// ...
  export interface IGetIngredientsRequest {
    readonly type: typeof GET_INGREDIENTS_REQUEST;
  }
  
  export interface IGetIngredientsSuccess {
    readonly type: typeof GET_INGREDIENTS_SUCCESS;
    payload: Array<IIngredient>;
  }
  
  export interface IGetIngredientsFailed {
    readonly type: typeof GET_INGREDIENTS_FAILED;
    readonly errorText: string;
  }

  export type TIngredientsAction =
  | IGetIngredientsRequest
  | IGetIngredientsSuccess
  | IGetIngredientsFailed;